AutoEP: drop the redundant float32 copy in the default combine (#8621)
# AutoEP: drop the redundant float32 copy in the default combine

## What

One line in `combine_from_routed`
(`deepspeed/module_inject/auto_ep_layer.py`), the default
`combine_impl="weighted_sum"` path:

```diff
-            output = (output.float() * top_scores.reshape(T, top_k, 1).float()).sum(dim=1).to(expert_output.dtype)
+            output = (output * top_scores.reshape(T, top_k, 1).float()).sum(dim=1).to(expert_output.dtype)
```
## Unblock an OOM issue

This line is where a large mixture-of-experts run stopped. Training
Qwen3.5-397B-A17B cut to 44
layers (about 292 B parameters for testing purpose) with ZeRO-3 +
AutoEP, expert parallel 32, 8,192 tokens per rank, on
4 nodes of 8 H200.

+++ deepspeed/module_inject/auto_ep_layer.py
@@ -455,7 +455,7 @@ def combine_from_routed(
         else:
             # Match the runtime HF grouped-mm path: apply routing weights per
             # token-slot sample, then reduce over top-k.
-            output = (output.float() * top_scores.reshape(T, top_k, 1).float()).sum(dim=1).to(expert_output.dtype)
+            output = (output * top_scores.reshape(T, top_k, 1).float()).sum(dim=1).to(expert_output.dtype)
     else:
         # Scores already applied pre-experts, just sum over top_k
         output = output.sum(dim=1)
